+++ CHANGELOG.md
@@ -9,5 +9,6 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Removed
 
 ### Fixed
+- Fix OpenSearch client to use refreshable credentials ([13](https://github.com/opensearch-project/opensearch-mcp-server-py/pull/13))
 
 ### Security